transport_api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
django==4.1.3
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via -r requirements.in
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via -r requirements.in
